Initial improvements to modularisation of codebase
