+++ lib/types.ts
@@ -1,10 +1,3 @@
-interface OnyxProps {
-  httpPathHandler: (path: string) => Promise<Response>;
-  s3PathHandler?: (path: string) => void;
-  fileWriter?: (path: string, content: string) => void;
-  extVersion?: string;
-}
-
 type ProjectField = {
   type: string;
   description: string;
@@ -19,4 +12,4 @@ type ResultType = Record<string, string | number | boolean | null>;
 
 type ErrorType = Record<string, string | string[]>;
 
-export type { OnyxProps, ProjectField, OptionType, ResultType, ErrorType };
+export type { ProjectField, OptionType, ResultType, ErrorType };
